HCI fixes
