+++ HCPInterface/cli/functions.py
@@ -18,77 +18,6 @@ from HCPInterface.hci import hci
 @click.command()
 @click.pass_obj
 @click.option('-q',"--query",help="Specific search query", default="")
-@click.option("-k", "--key",help="The path to the key-file of files to check for",default="")
-def check(ctx, query, key):
-    """Checks for file existence"""
-    if query != "":
-        try:
-            results= ctx["hci_query"]["results"]
-            for item in results:
-                itm = item["metadata"]
-                meta = itm["HCI_displayName"]
-                samples = itm["samples_Fastq_paths"]
-                string = "".join(samples).strip("[]").strip("{]}'")
-                lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
-            log.info("Metadata file: {}".format(meta))
-            for i in lst:
-                if query in i or query in os.path.basename(i):
-                    log.info("check: ",i)
-                    name = i.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
-
-        except:
-            log.info("File(s) does not exists: {}".format(query))
-
-    elif key != "":
-        obj = ctx["hcpm"].get_object(key) # Get object with key.
-
-@click.command()
-@click.option('-d',"--destination",help="Specify destination file to write to",required=True)
-@click.option('-l',"--legacy",help="Legacy mode to download files on the old NGS buckets",default=False,is_flag=True)
-@click.option('-q',"--query",help="Specific search query", default="")
-@click.option("-k", "--key",help="The path to the key-file on the HCP you want to download",default="")
-@click.pass_obj
-def download(ctx, destination, legacy, query, key):
-    """Download files using query for RunID or Sample name."""
-    if query != "":
-        results= ctx["hci_query"]["results"]
-
-        if legacy:
-            for item in results:
-                itm = item["metadata"]
-                samples = itm["samples_Fastq_paths"]
-            for i in samples:
-                obj = ctx["hcpm"].get_object(i) # Get object with json.
-                if obj is not None:
-                    ctx["hcpm"].download_file(obj, "{0}/{1}".format(destination,os.path.basename(i))) # Downloads file.
-                else:
-                    log.error("File: '{0}' does not exist in bucket '{1}' on the HCP".format(s,bucket))
-
-        elif not legacy:
-            for item in results:
-                itm = item["metadata"]
-                samples = itm["samples_Fastq_paths"]
-                string = "".join(samples).strip("[]").strip("{]}'")
-                lst = string.replace('"','').replace("\\","").replace("[","").replace("]","").replace(";",",").split(",")
-
-            for i in lst:
-                if query in os.path.basename(i) or query in i:
-                    s = os.path.basename(i)
-                    log.info("downloading:",s.replace(".fastq.gz", ".fasterq").strip())
-                    name = s.replace(".fastq.gz", ".fasterq").strip() # Replace suffix. 
-                    obj = ctx["hcpm"].get_object(name) # Get object with json.
-                    if obj is not None:
-                        ctx["hcpm"].download_file(obj, "{0}/{1}".format(destination,os.path.basename(name))) # Downloads file.
-                    else:
-                        log.error("File: '{0}' does not exist in bucket '{1}' on the HCP".format(name, bucket))
-
-    elif key != "":
-        obj = ctx["hcpm"].get_object(key) # Get object with key.
-        ctx["hcpm"].download_file(obj, destination) # Downloads file.
-
-@click.command()
-@click.pass_obj
-@click.option('-q',"--query",help="Specific search query", default="")
 @click.option("-k", "--key",help="The path to the key-file containing a list of queries",default="")
 def search(ctx, query, key):
     """Search for file"""
